*adjustment* no response send but still visible to users

+++ resources/js/composables/heartbeat.js
@@ -1,15 +1,17 @@
 import axios from "axios";
 
+const time = 1 * 60 * 1000 // 1 min
+
 export const startHeartbeat = (router) => { // Accept router here
     setInterval(async () => {
+        if (!window.location.origin) return; // Skip if origin is not set (e.g., during SSR or certain test environments)
         try {
-            const { data } = await axios.post('/api/heartbeat');
+            await axios.post('/api/heartbeat');
         } catch (err) {
-            console.log("Heartbeat caught an error:", err.response?.status); // Check this in Inspect > Console
             if (err.response?.status === 403 || err.response?.status === 401) {
                 localStorage.clear();
                 router.push('/');
             }
         }
-    }, 5000);
+    }, time);
 };
